requirements.txt
protobuf==7.34.1
psutil==7.2.2
python-dotenv==1.2.2
streamlit==1.55.0
\nweaviate-client==4.6.5\nPillow==10.3.0
